backend/requirements.txt
flask==2.1.0
flask-cors==3.0.10
werkzeug==2.1.2
mysql-connector-python==8.0.26